% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{fix_parameters}
\alias{fix_parameters}
\title{parameters which depend on others, etc}
\usage{
fix_parameters(y, Ncat, Nage, par_seq, condom_seq, groups_seq, years_seq)
}
\description{
Eugene's function to be documented
}
